+++ packages/engine/src/avatar/components/AvatarDissolveComponent.ts
@@ -23,21 +23,137 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { defineComponent } from '../../ecs/functions/ComponentFunctions'
-import { DissolveEffect } from '../DissolveEffect'
+import { Material, Mesh, MeshBasicMaterial, MeshStandardMaterial, ShaderMaterial } from 'three'
+import matches from 'ts-matches'
+import { Entity } from '../../ecs/classes/Entity'
+import { defineComponent, getComponent } from '../../ecs/functions/ComponentFunctions'
 
 export const AvatarDissolveComponent = defineComponent({
   name: 'AvatarDissolveComponent',
 
   onInit: (entity) => {
     return {
-      effect: null! as DissolveEffect
+      minHeight: 0,
+      maxHeight: 0,
+      currentTime: 0,
+      dissolveMaterials: [] as Material[]
     }
   },
 
   onSet: (entity, component, json) => {
     if (!json) return
 
-    if (json.effect) component.effect.set(json.effect as DissolveEffect)
+    if (matches.number.test(json.minHeight)) component.minHeight.set(json.minHeight)
+    if (matches.number.test(json.maxHeight)) component.maxHeight.set(json.maxHeight)
+    if (matches.number.test(json.currentTime)) component.maxHeight.set(json.currentTime)
+  },
+
+  createDissolveMaterial(object: Mesh<any, MeshBasicMaterial & ShaderMaterial>, entity: Entity): any {
+    const hasUV = object.geometry.hasAttribute('uv')
+    const material = object.material.clone()
+    const hasTexture = !!material.map
+
+    let fragmentShader = ''
+    let vertexShader = ''
+
+    const vertexNonUVShader = `
+    #include <fog_vertex>
+    vec2 clipSpace = gl_Position.xy / gl_Position.w;
+    vUv3 = clipSpace * 0.5 + 0.5;
+    vPosition = position.y;
+  `
+
+    const vertexUVShader = `
+    #include <fog_vertex>
+    vUv3 = uv;
+    vPosition = position.y;
+  `
+
+    const fragmentColorShader = `
+    #include <output_fragment>
+    float offset = vPosition - time;
+    if(offset > (-0.01 - rand(time) * 0.3)){
+    gl_FragColor = vec4(color.r, color.g, color.b, 1.0);
+    gl_FragColor.r = 0.0;
+    gl_FragColor.g = 1.0;
+    gl_FragColor.b = 0.0;
+    }
+    if(offset > 0.0){
+    discard;
+    }
+  `
+
+    let textureShader = `gl_FragColor = textureColor;`
+    if (hasTexture && (material as any).map.isVideoTexture) {
+      textureShader = `gl_FragColor = sRGBToLinear(textureColor);`
+    }
+
+    const fragmentTextureShader = `
+    #include <output_fragment>
+    float offset = vPosition - time;
+    vec4 textureColor = texture2D(origin_texture, vUv3);
+    ${textureShader}
+    if(offset > (-0.01 - rand(time) * 0.3)){
+    gl_FragColor.r = 0.0;
+    gl_FragColor.g = 1.0;
+    gl_FragColor.b = 0.0;
+    }
+    if(offset > 0.0){
+    discard;
+    }
+  `
+
+    const vertexHeaderShader = `
+    #include <clipping_planes_pars_vertex>
+    varying vec2 vUv3;
+    varying float vPosition;
+  `
+
+    const fragmentHeaderShader = `
+    #include <clipping_planes_pars_fragment>
+    uniform vec3 color;
+    varying vec2 vUv3;
+    varying float vPosition;
+    uniform float time;
+    uniform sampler2D texture_dissolve;
+    uniform sampler2D origin_texture;
+    float rand(float co) { return fract(sin(co*(91.3458)) * 47453.5453); }
+    vec4 sRGBToLinear( in vec4 value ) {
+      return vec4( mix( pow( value.rgb * 0.9478672986 + vec3( 0.0521327014 ), vec3( 2.4 ) ), value.rgb * 0.0773993808, vec3( lessThanEqual( value.rgb, vec3( 0.04045 ) ) ) ), value.a );
+    }
+  `
+
+    vertexShader = vertexShader.replace('#include <clipping_planes_pars_vertex>', vertexHeaderShader)
+    vertexShader = vertexShader.replace('#include <fog_vertex>', hasUV ? vertexUVShader : vertexNonUVShader)
+    fragmentShader = fragmentShader.replace('#include <clipping_planes_pars_fragment>', fragmentHeaderShader)
+    fragmentShader = fragmentShader.replace(
+      '#include <output_fragment>',
+      hasTexture ? fragmentTextureShader : fragmentColorShader
+    )
+
+    material.vertexShader = vertexShader
+    material.fragmentShader = fragmentShader
+
+    const myMaterial = Object.assign(new MeshStandardMaterial(), material)
+    myMaterial.onBeforeCompile = function (shader) {
+      shader.fragmentShader = fragmentShader
+      shader.vertexShader = vertexShader
+      myMaterial.uniforms.time = { value: 1 }
+    }
+    myMaterial.visible = material.visible
+
+    console.log(entity)
+    getComponent(entity, AvatarDissolveComponent).dissolveMaterials.push(myMaterial)
+
+    return myMaterial
+  },
+
+  updateDissolveEffect(entity: Entity, dt: number) {
+    const dissolveComponent = getComponent(entity, AvatarDissolveComponent)
+    dissolveComponent.currentTime += dt
+    for (let i = 0; i < dissolveComponent.dissolveMaterials.length; i++) {
+      ;(dissolveComponent.dissolveMaterials[i] as any).uniforms.time.value = dissolveComponent.currentTime
+    }
+    return dissolveComponent.currentTime >= dissolveComponent.maxHeight
   }
 })
