+++ packages/engine/src/avatar/systems/AvatarLoadingSystem.ts
@@ -62,7 +62,6 @@ import { VisibleComponent } from '../../scene/components/VisibleComponent'
 import { setupObject } from '../../scene/systems/SceneObjectSystem'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { TweenComponent } from '../../transform/components/TweenComponent'
-import { DissolveEffect } from '.././DissolveEffect'
 import { AvatarControllerComponent } from '.././components/AvatarControllerComponent'
 import { AvatarDissolveComponent } from '.././components/AvatarDissolveComponent'
 import { AvatarEffectComponent } from '.././components/AvatarEffectComponent'
@@ -187,8 +186,8 @@ const execute = () => {
             }
           }
           setComponent(entity, AvatarDissolveComponent, {
-            /** @todo refactor to not be just the first index */
-            effect: new DissolveEffect(avatarObjects[0], bbox.min.y / scale, bbox.max.y / scale)
+            minHeight: bbox.min.y / scale,
+            maxHeight: bbox.max.y / scale
           })
         })
     )
@@ -236,9 +235,7 @@ const execute = () => {
   }
 
   for (const entity of dissolveQuery()) {
-    const disolveEffect = getComponent(entity, AvatarDissolveComponent).effect
-
-    if (disolveEffect.update(delta)) {
+    if (AvatarDissolveComponent.updateDissolveEffect(entity, delta)) {
       removeComponent(entity, AvatarDissolveComponent)
       const effectComponent = getComponent(entity, AvatarEffectComponent)
       const avatarGroup = getOptionalComponent(effectComponent.sourceEntity, GroupComponent)
